issue #8528 Markdown links to Markdown pages with explicit page command are broken
In case a  `@page` command is the first line in a markdown file it is used as  "label" but also to construct the name of the output file and hence the regular markdown reference (like `[...](..)`) does not work anymore as here an internal label and filename is used  (`md_...`).
To overcome this problem the `@page <label> (title)` is changed into `@page md_<filename> (title)\ilinebr @anchor <label>`.