Incorrect C++ comment conversion due to {@code
When having:
```
/// \file
///
/// a java code section
/// {@code
/// Some text
/// }
/// and more
```
and a default Doxyfile, we get
```
cc.h:8: warning: reached end of file while inside a 'code' block!
The command that should end the block seems to be missing!
```
this is independent of the setting of `ENABLE_PREPROCESSING`.
It is caused as `{@code` is not recognized as a "special" comment  in this case, it is recognized when we are in C-type comment.