Specifying strings to be excluded from markdown processing
When having code like
```
/** @struct _MODEL_DICT_
 *  @brief This structure contains models.dat info
 *  @var _MODEL_DICT_::s_tech_class
 *  Member 's_tech_class' contains tech_class info
```
we get warnings like:
```
 warning: Member s_tech_class (variable) of struct _MODEL_SMALL_DICT_ is not documented.
```
even though the member is properly documented. This is due to the fact that the `_MODEL_SMALL_DICT_` is in the markdown processing seen as an italic string `MODEL_SMALL_DATA`.

To overcome this problem in a general way an new setting had to be introduces: `MARKDOWN_EXCLUDE`