Incorrect citations for entries with double dash and for cross references
When having a simple citation problem like:
```
/** \mainpage
- \\cite s--92 \cite s--92
- \\cite csos-frcms-95 \cite csos-frcms-95
*/
```
with the bib entries (a bit strange here but it is because of the example):
```
@phdthesis{s--92
, author =      "M. F. Sanner"
, title =       "??"
, school =      "UHA"
, address =     "France"
, year =        1992
, keywords =    "doctoral thesis"
, update =      "98.03 bibrelex"
}
@inproceedings{csos-frcms-95
, author =      "Michel F. Sanner and Arthur J. Olson and Jean-Claude Spehner"
, title =       "Fast and Robust Computation of Molecular Surfaces,"
, booktitle =   "Proc. 11th Annu. ACM Sympos. Comput. Geom."
, year =        1995
, pages =       "C60--C70"
, keywords =    "molecules, biology, visualization"
, crossref =    "s--92"
, update =      "98.03 bibrelex, 95.09 mitchell"
}
```
and the settings:
```
CITE_BIB_FILES  = mybib.bib
QUIET = YES
OUTPUT_DIRECTORY = docs
```
we get the warning like:
```
aa.h:2: warning: \cite command to 's' does not have an associated number
```

This is due to the fact that the `--` has been converted to `&ndash;`  in the cite IDs and the detection of the cite IDs was not prepared for this.
Also the output of the cross-references would contain the unknown / by doxygen not supported `<cite>` tag.