Redefinition of VHDL specifiers
Based on the simple VHDL code line:
```
library ieee;
```
we got the warning:
```
warning: Member ieeeieee (variable) of class filter_pkg is not documented.
```
note the double "ieee" and as "type" variable instead of library.

This problem lead to a "small" redesign of the specifiers definition into a `struct Spec`  (see `types.h`) as the specifiers were misused by VHDL and the current uint64 had no space to store all values anymore and a rename of the different Values (like `Entry::Published` into `SpecifierPublished` and `VhdlDocGen::RECORD` into `SpecifierRecord`, all names so that no name classes with `enum`s occurred).

The actual change for the double "ieee"  was done in the vhdlparser.jj (setting the args field to the empty field for library / use)
The problem regarding the usage of `variable` instead of `library` is corrected in memberdef.cpp (wrong name used for testing).

For VHDL Entity, Architecture, Package and PackageBody with a class definition the protection field was misused (leading to a strange occurrence of a +2 as the protection and e.g. VhdlDocGen::EntityClass and also the old VhdlDocGen::Entity were off by 2), this has been properly corrected by directly using the specifier.
Also the automatic setting of `EXTRACT_PRIVATE` and `EXTRACT_PACKAGE` for VHDL output are now no longer needed (which could also give strange effects, especially in multi language projects i.e. not really wanted output).

Furthermore a number of "simple" changes were necessary to overcome problems with `if` conditions etc.