Make email obfuscation CSP friendly
Obfuscation makes email addresses harder to harvest by inserting
into them HTML elements which are supposed to be invisible in the
rendered HTML (display: none). The style is specified inline, and
inline styles may be prohibited by content-security-policy HTTP
header, breaking the markup if generated html is served from the
server which sets such header. So move the style to CSS and refer
to it via `class` in HTML code to fix this.