issue #8355 Doxywizard 1.9.x doesn't start up properly
The problem here was that the "wizard/loadsetting" key in the registry is set (Using a set of settings defined by the user as a "default" set of settings) in an older version and that some of the keys didn't yet exist in that version but were required in the current version.
So we have to test whether or not the setting is valid and not a default "QVariant" is returned (as this will not have methods required by doxygen ./ there is no value and thus no update necessary and the current value can remain.

Note that the test contains of 2 parts:
- var.isDefault() i.e. whether or not the setting is present in the user "default" settings
- The name of the setting is known but not compiled into the current executable (e.g the SQLITE3 settings of the CLAN settings)

In the past this was, apparently, not necessary (or never noticed / reported).