bug_711041 Undocumented "enum" values in TOC of qhp files causes tag-mismatch and corrupt TOC.
The problem regarding the extra `</section> in the case of a `user` section in the doxyfile coul dbe solved by means of better calculating the "diff", though this lead to bad index sequences in the qhp navtree.
To solve the bad tree entries, for the enum values, a function `closeContentsItem` is introduced.
(in `layout_default.xml` some white space at the end of the line is removed.)