Correct reference text of a page reference
During the initial check for #9023 it was discovered that the patch  "issue #8528 Markdown links to Markdown pages with explicit page command are broken"  (i.e. #8896) changed the behavior of the text in the links (the link name was shown instead of the text of the link).
To overcome this problem a new internal only command `\ianchor` had to be introduced which also can contain the text of the link (for consistency reasons it is not possible to add an extra argument to the `\anchor` command.