Waning due to interpretation of non special '//' comment
When having a java example (of C example without preprocessing) of the form:
```
/** A marker class for configuration transitions that are defined in Starlark. */
public abstract class StarlarkTransition implements ConfigurationTransition {
  /**
   * Method
   */
  public Map validate() {
        // value as {@code maybeAliasSetting} above.
  }

  /**
   * tst {@code root}.
   */
  public ImmutableMap fie(){}
}
```
we get a warning like:
```
StarlarkTransition.java:14: warning: reached end of file while inside a 'code' block!
The command that should end the block seems to be missing!
```
due to the cat that an attempt is made to interpret the `{@code ...}` in non doxygen comment, this should not be done.