Issue #9075 .html Extension Removed from Tag File URLs
The problem encountered is due to the `.` in the directory part of  the tag file name entries like:
```
<filename>java.base/java/lang/Object.html</filename>
```
during the processing doxygen strips the extension when it matches the extension of the current project (`HTML_FILE_EXTENSION`), later on doxygen attempts to check whether there is an extension present and when not to add the current extension.
The check is presotrmed on `java.base/java/lang/Object` and here an extension is "found" (`.base/java/lang/Object`), doxygen should here only consider the file name part.