issue #9070 Doxygen is not showing all called functions in graph plot
The end of an IDL cpp_quote was not properly caught (the rule `<*>\"` was expected to be used (Note: later on a rule dedicated to PHP was present but could / should not be reached, joined these 2 as well).
For the state `DefineEnd` there was already a separate rule, adding one for functions as well (until now the rule with `.` kicked in).