bug_644899 New diagram implementing enviromen
In principle this should already be possible by means of defining the relevant `EXTRA_PACKAGES` and embedding the LaTeX code in the right environment.
Due to the `simpliftWhiteSpace` though embedded `\n` were stripped as well so the entire LaTeX code was seen as one line and when having comment signs (`%`) in it, the rest of the line was not handled by LaTeX.
The formulas are a more verbatim type of so has to be used "as is".