HTMLHELP warning on using svg as image format for graphs
In the chm file, when having (interactive) SVG as format for the dot graphs we miss the `svgpan.js`. All used style sheets are added to the chm file.
When the `GENERATE_LATEX=YES` also the generated pdf files were added to the list of files, giving warnings (only files generated for the HTML version are relevant here).