Automatically set right executable properties for windows
In case the `editbin` executable is available automatically set the right properties  of the created executables.
For de bugging purposes it might be useful to have the `doxywizard` not as a "windows" type of executable but as "console" executable, this can be accomplished by means of the setting of `enable_console=ON`.
The `enable_console` only operates on the `doxywizard` executable. Usage like: `cmake -Denable_console=ON ..`