Show date in a customized form
Show a specified or the current date in a specified format.
This format translates also the name of the weekday and the name of the month to the names used according to the `OUTPUT_LANGUAGE` setting.
When the `<data>` is not specified the current date is used and this can be used for e.g. automatically adjusting the copyright message in the documentation (see e.g. `index.doc`)

(The names are derived by means of a python script from the python distribution).