issue #9110 MULTILINE_CPP_IS_BRIEF breaks multi-line latex formulas
In normal case the multiline `///` comments are replaced by comment blocks of type `/** ... */` an in here the extra `*` on a line are problematic with some commands. In case of `MULTILINE_CPP_IS_BRIEF=YES` this comment conversion is not done but in e.g. formulas the `///` was still stripped. This should not be the case.