Better handling of of *. in C++ ///  comments
In case we have a cpp comment like:
```
///     double "somepath/bin*/exename[.ext]", the resulting
///
```
this results in the (HTML) output like:
```
double "somepath/bin*&zwj;/exename[.ext]", the resulting
```
as the `*/` has bee converted in the comment conversion to `*&zwj;/` which is done due to the conversion of the `///` type comment to `/*` type comment.
In the string this is later seen as a literal `&zwj;` tjhat should remain.

To overcome the problem the replacement is done with a new doxygen internal entity `&doxyzwj;` that is filtered out again later on.