Downgrade stale translator warning to message
If the WARN_AS_ERROR configuration option is set to YES or
FAIL_ON_WARNINGS, a stale translator would cause the documentation
generation to fail (or at least to generate a non-zero return code),
even though there is nothing wrong with the documentation inputs. As
this is undesired, downgrade the message from a warning to only a
message.