Calculation of the reference Indentation during markdown processing
When having a simple comment like:
```
/// \file

/// Initial text
///
/// \verbatim
/// V1
/// \endverbatim
///
/// \htmlonly
/// H
/// \endhtmlonly
///
void fie1();
```
the word  `\htmlonly` will also be shown in a "verbatim" way.
This is due to the fact that the `detab` function sees the remaining `///` in the "verbatim"  type of blocks also and decides that the line starts at the beginning ogf the line and resets the reference indentation value.
The reference indentation should not be influenced by "verbatim" blocks.