Don't interpret */ in "verbatim" blocks
When having:
```
/// \verbatim
/// verbatim bare somepath/bin*/exename[.ext], the resulting
/// \endverbatim
/// \htmlonly
/// htmlonly bare somepath/bin*/exename[.ext], the resulting<br>
/// \endhtmlonly
```

the "end of  comment" `*/` is ignored in the `\verbatim` block but not in the `\htmlonly` block and resulting in warnings like:
```
warning: found */ without matching /* while inside a \htmlonly block! Perhaps a missing \endhtmlonly?
```
The `*/` should not only be ignored in `\verbatim` blocks but in all "verbatim" blocks.