Detecting end of body in python
When having:
```
def parsexmlstring0():
    return element

try:
    from generatedsnamespaces
except ModulenotfoundExp_ :
    GenerateDSNamespaceDefs_ = 0

def parsexmlstring1():
    return element

if (true):
    from generatedsnamespaces
else:
    GenerateDSNamespaceDefs_ = 1
```

we see that in the inline sources also the lines with the `try` and `from`. The line with `if` is not shown.
Similar exercises with showing of e.g `print(42)` or `x=42`.

The cause of the problem is that only words followed by a blank were recognized and not words with other characters.