Don't create anchor for \name command
Do not create an anchor for a `\name` command,  the same `\name` can be used multiple times in the same file / documentation block (most notably the `\name` without a text) leading to multiple anchors with the same values.
Furthermore the anchor is internally generated by doxygen, so the user has no, natural, way to reference it.

The function `visibleMembergGroup` is not used anywheer, so obsolete and thus removed.