Handling enums in settings that double as boolean
As found by CGAL the `CLASS_GRAPH` setting doubles nowadays as enum (to determine to show the class graph as text / graph / both or none) and as boolean (for the visible attribute of the layout file).
Due to the changes of #9278 the values "TEXT" and "GRAPH" were not handled correctly.