Error when generating documentation with formulas a second time
With the pull request #9088 the embedded `\n` are left in the formula and as formulas now can span multiple lines some lookahead is necessary.

The problem shows when doing a documentation run multiple times and a formula spans multiple lines (and MathJax isn't used). The problem surfaced when generating the doxygen manual (due to changes)  a second time.