Implementation of lineBreak and nonBreakingSpace in docbook
In the docbook it is hard to see and implement a lineBreak or a nonBreakingSpace.

The advise from the documentation for the nonBreakingSpace is to use either `&#xA0;` or `&nbsp;` (see: http://www.sagehill.net/docbookxsl/SpecialChars.html) and as we used as replacement for the HTML `&nbsp;` already `&#xA0;` it is used here as well.

The advise from the documentation for the lineBreak is to use either `<literallayout >` or to use <`?linebreak?>` (see: http://www.sagehill.net/docbookxsl/LineBreaks.html) where the `<literallayout>`  is not suitable in all cases..
The `<?linebreak?>` gives the user of the docbook more flexibility as:

> You can create an XML processing instruction to indicate a line break, and add a stylesheet customization to enact it. A template like the following can be added to a customization layer:
>
> ```
>   <xsl:template match="processing-instruction('linebreak')">
>     <fo:block/>
>   </xsl:template>
> ```

This method is now also implemented for the replacement of the `<br>` tag.