issue #9016 Support collapsible sections in markdown
With #8890 the HTML details tag had been implemented as a style change though this give a problem when inside the details part there are multiple paragraphs.
The HTML details tag has been reimplementd as a so called `DocNode`, so that it is possible to have multiple paragraphs in a detailed section.