issue #9344 Long bibliography entry looks bad
When having the citation entry (here an artificial example):
```
@inproceedings{key,
 author = {Author},
 title = {Title},
 booktitle = {Proceedings of the 24th annual conference on publishing proceedings from very interesting conferences},
 series = {Conference '22},
 year = {2022},
 pages = {123--132},
 numpages = {10},
 publisher = {Publisher},
 address = {Address},
}
```
this results in a `bbl` file entry like:
```
<dt><a name="CITEREF_key">[Author, 2022]</a></dt><dd>\bibxhtmlname{Author}.
Title.
In <em>Proceedings of the 24th annual conference on publishing proceedings from
  very interesting conferences</em>, Conference '22, pages 123--132, Address,
  2022. Publisher.</dd>
```
and thus in a `citlists.doc` entry:
```
<!-- Authors: Author -->
<dt><a name="CITEREF_key">[2]</a></dt><dd>Author.
Title.
In <em>Proceedings of the 24th annual conference on publishing proceedings from
  very interesting conferences</em>, Conference '22, pages 123&ndash;132, Address,
  2022. Publisher.</dd>

</dl>
```
and due to the number at the beginning of the line this is interpreted, by doxygen, as a list start and resulting in:
```
1. Publisher
```

We have to see to it that "continuation" lines from the bbl file (i.e. new lines starting with `  ` are joined to the previous lines and as the complete entry is read in one got it is sufficient to replace `\n  ` by ` `.