Enabling include files for lex files
The motivation for this is that make the rules and code as used in the different lexers more consistent as now some definitions are used in multiple lexers (e.g. for white space) and they are not always consistent.
This patch provides the infrastructure so that late on it is easy to add the include files.

The (f)lex program is missing an include facility and the suggestion (https://stackoverflow.com/questions/66197941/defining-lex-include-files-for-flexers) was to write an own pre-processor.
The code presented here functions as a pre-processor and consists of a number of states:
- pre processing the `.l` file to include files and remember where these lines are included
- creating the so called scan states include file, this was already present but now uses the result of the previous step
- translate the pre-processed `.l` file to an `.cpp` file
- post-process the file created `.cpp` file from  the previous step so that line numbers and used files refer to the original file(s) and the now adjusted `cpp` file.