Incorrect determination of sub-sections for index.
Having an example like:
```
/*! \page commands Special Commands

\section cmd_intro Introduction

\section cmd_intro1 Introduction1

\anchor anc1

\section cmd_intro2 Introduction2
*/
```
the sections `Introduction` and `Introduction2` are shown as "pages" in the index of the CHM file but the section `Introduction1` is shown as a "book".

The problem is the `anchor` in the `Introduction1` section as an `anchor` (and `table`) are also handled with section references but should have been excluded here as it was done earlier for the current section for which the sub-sections are determined.