No click possible on chapters in doxygen's CHM documentation
In the pull request #8843 (for the issue #6000  Class Members is missing in the CHM. (Origin: bugzilla 766164)) a fix was made for the fact that the `Classes` item was not shown in the CHM index.
The side effect of this PR is that now the main "books" / chapters cannot opened anymore which can be seen in e.g. the CHM  version of the doxygen manual.
E.g. when one wants to see the introduction text in the chapter `Markdown support ` one first has to decent into the tree:
`Markdown support ` -> `Standard Markdown` -> `Paragraphs` and at that moment it is possible to scroll up to the top of the page.

The problem in the original problem (bug 766164) was that there are 2 consecutive items with a different name but with the same reference (here to "annotated.html").
The problem only occurs at the moment that an "page" and a "sub page" refer to the same generated page, by comparing the references to have the same file name and an empty anchor we know we are in this situation and as we are at page level we also know that an `id=top` is present so for the second reference (first one has already been completely handled) we can use this `id` to get the references correct and the doxygen CHM manual works again and also the bug 766164 is gone.