Dependency between examples.
When building the `tag` example a prerequisite is that the `example` example has been build (as the `tag` example needs the `example.tag` file, so we have first generate the `example` example before building the `tag` example).
An explicit dependency structure for relations between examples has been defined.

(based on the problem as reported in the proposed patch:  delete TAGFILE for build error: #9387 )