issue #9399 XML / perlmod fortran output, the type of a parameter of a function is set to the name of the parameter
Fortran has, regarding the definition of the type of an argument a bit of a different style compared to other languages, i.e. the type is defined, more or less, defined in the body and by doxygen not stored in the declaration list but only in the definition list.

Added for Fortran the possibility that the type should come from the definition list (when available).