Files Quick Link results in a non existing page
When investigating bug 771822 ( Doxygen wrongly infers that a macro invocation in a header file is a function rather than a variable declaration (Origin: bugzilla 771822) #6076 ) in the current version there the quick link "Files" was shown although there was the setting `SHOW_FILES=NO`. (

The problem was not present in the 1.9.3 version but appeared in the 1.9.4 version. A bisection revealed:
```
7f514f1822de413da25bd5a6eba8938a3483e793 is the first bad commit
commit 7f514f1822de413da25bd5a6eba8938a3483e793
Date:   Sun Apr 24 13:35:58 2022 +0200

    SHOW_FILES=NO could cause broken links for grouped files

 src/index.cpp | 44 ++++++++++++++++++++------------------------
 1 file changed, 20 insertions(+), 24 deletions(-)
```
due to the removal of the test on `SHOW_FILES` here it has to be added in the quick link section test.