Prohibit in-place doxygen builds
Based on some build problems in issue #9413,  it is checked that there is no attempt to build doxygen in-place (i.e. that the main source directory and the build directory have to be different directories).
(The original problem shows errors like:
```
cycle in dependency tree for target 'generated_src\configimpl.corr'
```
)