Prohibit in-place doxygen builds
With `make` it looks like the warning about the cyclic dependency does not result in a stop of the build like when using `nmake`.
The problem is introduce by the requirement of the coverage tool of gcc to have the sources in the build tree (see comment in the `src/CMakeLists.txt`).
The copying and the `FATAL_ERROR` is now restricted to the case where the coverage is enabled.