issue #9440: Fix search label regression
== Background:

The extension value (`.html`) was injected instead of the search
label.

This meant that upon focussing the field it was no longer cleared,
so the "Search" text remains in the input field and thus had
to be removed manually.

It also meant that after de-focussing the field (click anywhere),
the field value was assigned to `.html` which made it show up in the
user interface even.

== Change:

Use the HTML5 placeholder attribute instead, which is a browser
built-in that does this automatically. It displays the value
while the field is empty, and makes it dissappear when the empty
field is focussed.

This has the added benefit of no longer destroying the user input
if you type something and then click or tab elsewhere before
submitting the form.

== Also:

* Remove the now-unused code for managing this value.
* Synchronise the signature of the two SearchBox js classes to avoid
  mistakes like this in the future. The `inFrame` argument is not
  used in extsearch.js. Replace it with an equally-unused `extension`
  arg for consistency with search.js.

Fixes https://github.com/doxygen/doxygen/issues/9440.