[regression] Non resolved ALIASES with minus sign
Small regression on  #9451 issue #9047 Support @param-taint in PHP (permit dashes in command aliases)?
In a number of project (e.g. CGAL) we now got the warning like
```
warning: Found unknown command '\cgal'
```
as the code was:
```
\cgal-like
```
and
```
ALIASES += cgal=%CGAL
```

made the alias replacement so that when the expansion didn't do anything and the input string contains a `-` sign the last `-` sign is stripped as well as the part after it and a new attempt is made to resolve the command.