bug 786576 regression with ALIASES or image latex interpretation
The problem is not the ALIAS but just the way the `\image` command handles its (complex) set of arguments.
For this reason it is also advised to use `^^` i.e. an artificial line break in the `ALIASES` after a `\image` command.

The problem is that at with the image command and argument was "read" and checked, but when it didn't fulfill the requirements it was discarded   (and thus lost).

The arguments of the `\image` command are well defined and we can see that there won't be commands (i.e. starting with `\` or `@` like `\htmlonly`) , symbols (starting with `&` like `&nbsp`) or html tags  (starting with `<`).
These constructions are, when found, terminating the `\image` command and pushed back into the "read stream".