Recognizing and name of implicit Fortran [programs
In Fortran it is possible that a part of code doesn't start with a `program` / `subroutine`/ `function` etc statement and that is should be considered as the main program.
Some examples have multiple sections with "main" programs (this won't compile, but serves just as example code.
When this happens we see warnings like:
```
.../music5.f:693: warning: documented symbol 'program __music5 f__' was not declared or defined.
.../music5.f:985: warning: documented symbol 'program __music5 f__' was not declared or defined.
```
In the resulting code we see a number of problems
- just one "main" program
- the name in the warning containing a space

after correcting this we still see in the source code
- incorrect reference to begin of the "main" program as not all constructs are recognized as starting a "main" program.