issue #9502 Fortran containing a single apostrophe (') in comments interferes with Doxygen parsing.
In the preprocessor we have to ignore Fortran comment not only at the beginning of the line but also after a statement like `INTEGER :: I !< The user's comment`.
Furthermore in Fortran fixed format code also lines that have `C`, `c` or `*` at the first position should be skipped (see also util.cpp and commentcnv.l).