Full line environment replacement for configuration setting
With the pull request #9509 the possibilities for the  CMake replacement variables as part of a setting was added, though the full line replacement was not possible, giving warnings like:
```
warning: ignoring unknown character '@'
warning: ignoring unknown character '$'
```
this has now been corrected.
This feature is useful when one has e.g.
```
  if (DOXYGEN_VERSION VERSION_LESS 1.9.3)
    set(GUDHI_DOXYGEN_CLASS_DIAGRAMS "CLASS_DIAGRAMS = NO")
  else()
    set(GUDHI_DOXYGEN_CLASS_DIAGRAMS "")
  endif()
```