Added translation specialization for VHDL in the `trMemberFunctionDocumentation()` method, (see method body with `OPTIMIZE_OUTPUT_VHDL` condition in the `translator_en.h` file). This VHDL specialization string `Member Function/Procedure/Process Documentation` is based on the translation string in the `trFunctionAndProc()` method.
Added the possibility to translate the `flowchart: ` string in `htmldocvisitor.cpp` and todo remark describing this here was removed.

Appended translation section 1.9.6 to the `translator_en.h` file.
Flowchart string added as new method `trFlowchart()` to this section and string was changed to: `Flowchart: `.

Added `class TranslatorAdapter_1_9_5` to the file `translator_adapter.h`.
Changed dependency of up-to-date translations: `br`, `cz`, `nl`, `pt` to `TranslatorAdapter_1_9_5`.