Hash comment in php
The hash (`#`) comment was not recognized in the comment converter for php so sthe code like:
```
<?php
## page's elements with
class PageStore {
  function ls($pats=NULL) {
    $dirlist = array(preg_replace('!/*[^/]*\\$.*$!','',$dir));
    return $out;
  }
}
```
resulted, incorrectly, in the warning
```
aa.php:9: warning: Reached end of file while still inside a (nested) comment. Nesting level 1 (probable line reference: 5)
```
as the single quote in the comment was interpreted and this should not have been the case.