Warning during compilation on Windows systems
This is based on the warning type: C4800: Implicit conversion from 'type' to bool. Possible information loss
A real bug was found in the writing out of the boolean replacement values in configgen.py. The boolean values got double quotes and the string would always result in true instead of the intended value