issue #9584 Warning "Illegal command \ifile found as part of a \c command" in \copydoc, but not original documentation
The (internal command `\ifile`, for better warnings ) was placed directly after the existing pext an so the `\ifile` was part of of the `nullptr` text.
Problem occurred only when no other text followed.
We need a `<space>` as well as the, internal, command  `\ilinebr` e.g. for the `\addindex` command as this would run to the end of the line and would with a space still show the same problems.