Usage of steady_clock
Use the steady_clock (like in portable.cpp and doxygen.cpp) as the system_clock can be adjusted (e.g. leap seconds, daylit saving time, user adjusting the clock on the computer).

**steady clock**
Objects of class steady_clock represent clocks for which values of time_point never decrease as physical
time advances and for which values of time_point advance at a steady rate relative to real time. That is,
the clock may not be adjusted.

**system_clock**
Objects of type system_clock represent wall clock time from the system-wide realtime clock.