issue #9615 doxygen C# regression
The problem is that in case non of the output generators HTML, LaTeX, RTF, docbook or Man are used not all the settings are done before the tests are started.
We see this problem also when we just set `GENERATE_XML`, we get the warning like"
```
Main.cs(21): warning: parameters of member CSharpTutorials.Program.Main are not documented
```

by introducing a new output generator that will only be enabled in case non of the generators HTML, LaTeX, RTF, docbook or Man is enabled, the settings are done, so the tests can be executed properly.