Fix unreadable file warnings for broken symlinks with EXCLUDE_SYMLINKS is set
When walking directories to search for source files, Doxygen prints a
warning for each file that isn't readable. This can be caused by broken
symlinks. When EXCLUDE_SYMLINKS is set to YES, symlinks are skipped
anyways, so there is little point in printing a warning. Refactor the
readDir() and readFileOrDirectory() functions to avoid the warning in
that case.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>