Create possibility to disable a user defined xref list
There is a discrepancy between the standard defined xref lists (bug, todo, test, deprecated) and the user defined defined xref list, as there is no, easy, equivalent for a command like `GENERATE_BUGLIST=NO`.
Defining the user defined xref list using the `\noop` doesn't work as here only the first line is removed and not the rest of the lies of a paragraph.
Only solution is to have the user defined xref list item embedded in `\if ... endif` and use `ENABLED_SECTIONS`, but this is just a workaround.
Introducing the `GENERATE_XREFLIST` creates the possibility to make a user and standard defined xref list behave in the same way.