Consistency between HTML and LaTeX / RTF for examples
In the treeview of the HTML output we see the text `Examples` and also at other places,  though in the the LaTeX and RTF output we see `Example Documentation`.
To make all consistent everythin has been made consistent by means of using the function `trExamples` (and removing the now obsolete `trExampleDocumentation`.