Handling of text for the `\ref`  command
The, optional, text for a `\ref` command has to be, according to the documentation, between double quotes and the text should not s[an lines:
> \ref <name> ["(text)"]

the implementation is so that the text ends either with a double quote or an end of line.
When the text runs over the line into the next line we have on the next line just the beginning of a string starting with a double quote.
The chosen solution is so that the text now ends at at the closing double quote and that the end of lines are allowed (they are later on seen as whitespace).