Incorrect man code output in respect to backslashes
When having a simple  code section in the documentation like:
```
/**
 The fie
 \code
 #define X \
   Y
 char *b = "Some \fB string.";
 \endcode
*/
void fie();
```
and looking at the man output we see that the backslashes (`\`) are interpreted by the man interpreter as doxygen din;t properly escape them, resulting in
```
       The fie

       #define X   Y
       char *b = "Some  string.";
```
i.e. joining the `#define` lines and printing the word `string and the rest of the section in bold.