issue #9223 take param/return docu in account from the header file when generating docu for the source file
Based on bug 163033 Usability issues caused by single detailed comment requirement (Origin: bugzilla
a new looks was thrown at this issue especially in respect to the different possibilities of having an not having brief and detailed descriptions in the `.h` and `.cpp` files.
The output is now made consistent so that the information is always present and in the same order.
Special attention has been payed to the line counting.