issue #9733 doxygen stops parsing a file if one uses latex formulas and //!< or ///< for inline parameter documentation
When copying the data of a formula the formula is ended with the corresponding `\f` command. It makes no real sense to check here for other characters (like `[A-Za-z\-]`) as the formula is terminated with a special end sequence, contrary to other verbatim blocks where we have more general words.